+++ views/partials/social-button.php
@@ -1,0 +1,3 @@
+<a class="social-button social-<?= $medium; ?>" href="<?= $link; ?>" target="_blank" title="<?= $title; ?>">
+    <i class="icon-<?= $medium ?>"></i>
+</a>